requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.9.0
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
bleach==6.2.0
Brotli==1.1.0
bs4==0.0.2
cachetools==5.5.2
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.5
cssselect2==0.8.0
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django==4.2.23
django-anymail==13.0
django-cors-headers==4.7.0
django-markdownify==0.9.5
django-redis==6.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
executing==2.2.0
fonttools==4.59.0
frozenlist==1.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
humanize==4.12.3
idna==3.10
ipython==8.37.0
jedi==0.19.2
jiter==0.10.0
kombu==5.5.4
Markdown==3.8.2
matplotlib-inline==0.1.7
multidict==6.6.3
oauthlib==3.3.1
openai==0.27.10
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
playwright==1.54.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python3-openid==3.2.0
redis==6.2.0
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.7.0
sorl-thumbnail==12.11.0
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
tabulate==0.9.0
tiktoken==0.3.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
yarl==1.20.1
zopfli==0.2.3.post1
